Remove transforms demeaning in RandomMotion
Sorting the spectra so that the original image is in the middle of the k-space
has a similar effect to demeaning all the transforms, therefore is a bad idea
to use both. This commit remove the demeaning for now. We might implement the
exact method of Shaw et al. in the near future.

Resolves #85.