Deep-copy sample before applying transform
The current behavior is inconsistent across different transforms. Some of them
happen in-place, modifying the input, and for others, the input is kept as it
was. As discussed in #26, the input should remain unchanged.

The copy.deepcopy function seems fast enough and ensures that the input will
not be modified after applying a transform to it.

Resolves #26.