Add kwargs to Image and Subject
This commit allows the user to easily add values to samples generated by
ImagesDataset and to the image dictionaries in the samples.

This is especially useful for regression tasks of a single value, which
can be added to the subject as `Subject(image1, image2, label=1)` or
the image as `Image(name, path, type_, label=1)`.

Resolves #112.

Thanks @nwschurink for bringing this up.