Use slicing and clone() to crop images in samplers
Using sitk.Crop is slow because sitk.GetImageFromArray is slow.
Using copy.deepcopy(patch) is slow because it copies the whole array,
not just the patch. The solution is to use slicing for cropping and
patch.clone() for copying.

Related to #175, #183, 89acf631d3329c83ec91ac526ebe7ae7ae433cd6.
Fixes #212.