Fix plotting a subject with a single image
The figure and axes are created using `plt.subplots(r, c)`, where `r`
is the number of rows and `c` is the number of columns, typically 3
for a medical image. If r = 1, the array of axes will be 1D.

Each row of this array was being passed to `plot_volume`, which then
indexed it from 0 to 2, for the three orhogonal views. This was raising
a `TypeError`.

Reported in
https://github.com/fepegar/torchio/issues/336#issuecomment-718481336