fixed typo in quickstart
Hello,

I've just started using torchio and used quickstart as example to the project I want to start.
I've encountered a error:

```
NameError                                 Traceback (most recent call last)
<ipython-input-5-df67c26faa17> in <module>
     20 # apply RandomAffine 80% of the times and RandomElasticDeformation the rest
     21 # Also, there is a 25% chance that none of them will be applied
---> 22 spatial = OneOf({
     23         tio.RandomAffine(): 0.8,
     24         tio.RandomElasticDeformation(): 0.2,

NameError: name 'OneOf' is not defined
```

As you can see, quickstart guide gives error saying that `OneOf` is not defined, and it's not imported anywhere, so I was like "Hmm... Is it torch, albumentation or torchio implementation?". Then I looked at transofrmation and found that there is an implementation of `OneOf`, so I simply added `tio.` to `OneOf` and it worked.

With this PR I just want it to be fixed for everyone so that they will not have to spend extra time looking for their answer (which `OneOf` is that). Also it'll be helpful for starters to have working solution :)