Remove channels_last kwarg from Image
Fixes #684.

SimpleITK is not very good for fMRI and dMRI images.
After this commit, if 4 spatial dimensions are detected, we assume that
the read image is of this nature and we move convert the temporal
dimension into the channels dimension.

The `channels_last` kwarg should no longer be necessary.

It is important to take into account that images saved by TorchIO
always put the channels in the 5th NIfTI dimension, not the 4th.